♻️ refactor editor redux code into smaller files
this also switches away from redux-toolkits createSlice due to its usage making code-splitting hard.
