+++ app/src/pages/Editor/SelectionMenu.tsx
@@ -1,11 +1,12 @@
 import { useDispatch } from 'react-redux';
-import { exportSelection, copySelectionText } from '../../state/editor';
 import * as React from 'react';
 import { HTMLAttributes, RefObject, useRef, useState } from 'react';
 import styled, { css } from 'styled-components';
 import { useDocumentEvent } from '../../components/useListener';
 import { Button, Group } from 'evergreen-ui';
 import { useElementSize } from '../../components/useElementSize';
+import { exportSelection } from '../../state/editor/io';
+import { copySelectionText } from '../../state/editor/edit';
 const SelectionMenuContainer = styled.div<{ noMouse: boolean; centerHorizontally: boolean }>`
   position: absolute;
   display: block;
