+++ app/src/state/index.ts
@@ -21,4 +21,3 @@ export const store = configureStore({
 (window as any).store = store;
 
 export type RootState = ReturnType<typeof store.getState>;
-export type AppDispatch = typeof store.dispatch;
